extract FormDialog from Shift and Qualification dialogs
