--- src/ui/shift-dialog/styles.module.css
@@ -1,9 +1,0 @@
-.fullScreenDialog {
-  position: absolute;
-  top: 0;
-  left: 0;
-  bottom: 0;
-  right: 0;
-  border-radius: 0;
-  background: var(--color-background);
-}
